[chores] Update docs, refactoring, update sample app
