+++ docs/index.rst
@@ -48,8 +48,8 @@ the OpenWISP architecture.
     user/zerotier.rst
     user/openvpn.rst
     user/subnet-division-rules.rst
+    user/whois.rst
     user/rest-api.rst
-    user/who-is.rst
     user/settings.rst
 
 .. toctree::
