+++ openwisp_controller/config/base/whois.py
@@ -8,26 +8,33 @@ from jsonfield import JSONField
 
 from openwisp_utils.base import TimeStampedEditableModel
 
-from ..whois.service import WhoIsService
+from ..whois.service import WHOISService
+from ..whois.tasks import delete_whois_record
 
 
-class AbstractWhoIsInfo(TimeStampedEditableModel):
+class AbstractWHOISInfo(TimeStampedEditableModel):
     """
-    Abstract model to store WhoIs information
+    Abstract model to store WHOIS information
     for a device.
     """
 
     id = None
+    # Using ip_address as primary key to avoid redundant lookups
+    # and storage of duplicate WHOIS information per IP address.
+    # Whenever a device's last ip address changes, data related to
+    # previous IP address is deleted. If any device still has the
+    # previous IP address, they will trigger the lookup again
+    # ensuring latest WHOIS information is always available.
     ip_address = models.GenericIPAddressField(db_index=True, primary_key=True)
     isp = models.CharField(
         max_length=100,
         blank=True,
-        help_text=_("Organization associated with registered Autonomous System Number"),
+        help_text=_("Organization for ASN"),
     )
     asn = models.CharField(
         max_length=6,
         blank=True,
-        help_text=_("Autonomous System Number"),
+        help_text=_("ASN"),
     )
     timezone = models.CharField(
         max_length=35,
@@ -53,7 +60,7 @@ class AbstractWhoIsInfo(TimeStampedEditableModel):
             raise ValidationError(
                 {
                     "ip_address": _(
-                        "WhoIs information cannot be created for private IP addresses."
+                        "WHOIS information cannot be created for private IP addresses."
                     )
                 }
             )
@@ -72,16 +79,14 @@ class AbstractWhoIsInfo(TimeStampedEditableModel):
     @staticmethod
     def device_whois_info_delete_handler(instance, **kwargs):
         """
-        Delete WhoIs information for a device when the last IP address is removed or
+        Delete WHOIS information for a device when the last IP address is removed or
         when device is deleted.
         """
         if instance._get_organization__config_settings().whois_enabled:
-            transaction.on_commit(
-                lambda: WhoIsService.delete_whois_record.delay(instance.last_ip)
-            )
+            transaction.on_commit(lambda: delete_whois_record.delay(instance.last_ip))
 
     # this method is kept here instead of in OrganizationConfigSettings because
-    # currently the caching is used only for WhoIs feature
+    # currently the caching is used only for WHOIS feature
     @staticmethod
     def invalidate_org_settings_cache(instance, **kwargs):
         """
@@ -89,7 +94,7 @@ class AbstractWhoIsInfo(TimeStampedEditableModel):
         Organization settings instance.
         """
         org_id = instance.organization_id
-        cache.delete(WhoIsService.get_cache_key(org_id))
+        cache.delete(WHOISService.get_cache_key(org_id))
 
     @property
     def formatted_address(self):
