+++ openwisp_controller/config/whois/handlers.py
@@ -5,27 +5,25 @@ from .. import settings as app_settings
 
 
 def connect_whois_handlers():
-    # if WHOIS_CONFIGURED is False, we do not connect the handlers
-    # and return early
     if not app_settings.WHOIS_CONFIGURED:
         return
 
     Device = load_model("config", "Device")
-    WhoIsInfo = load_model("config", "WhoIsInfo")
+    WHOISInfo = load_model("config", "WHOISInfo")
     OrganizationConfigSettings = load_model("config", "OrganizationConfigSettings")
 
     post_delete.connect(
-        WhoIsInfo.device_whois_info_delete_handler,
+        WHOISInfo.device_whois_info_delete_handler,
         sender=Device,
         dispatch_uid="device.delete_whois_info",
     )
     post_save.connect(
-        WhoIsInfo.invalidate_org_settings_cache,
+        WHOISInfo.invalidate_org_settings_cache,
         sender=OrganizationConfigSettings,
         dispatch_uid="invalidate_org_config_cache_on_org_config_save",
     )
     post_delete.connect(
-        WhoIsInfo.invalidate_org_settings_cache,
+        WHOISInfo.invalidate_org_settings_cache,
         sender=OrganizationConfigSettings,
         dispatch_uid="invalidate_org_config_cache_on_org_config_delete",
     )
